Search by slug or id implemented

+++ ads/contollers/cars/carController.js
@@ -1,3 +1,4 @@
+const mongoose = require('mongoose');
 const RequestIp = require('@supercharge/request-ip');
 const Car = require('../../models/cars/carModel');
 const CarView = require('../../models/cars/car-views/ip-views-model');
@@ -126,10 +127,26 @@ exports.getAll = catchAsync(async (req, res, next) => {
 });
 
 exports.getOne = catchAsync(async (req, res, next) => {
-  const result = await Car.findById(req.params.id).populate('createdBy');
+  let result;
+  const ObjectId = mongoose.isValidObjectId(req.params.id);
+  if (ObjectId !== true) {
+    let stringValues = req.params.id;
+    let splitValues = stringValues.split('-');
+    // let idFromValues = splitValues.pop();
+    let idFromValues = splitValues.slice(-1)[0];
+    req.params.id = idFromValues;
+    let validId = mongoose.isValidObjectId(req.params.id);
+    if (validId === true) {
+      result = await Car.findById(req.params.id).populate('createdBy');
+    }
+  } else {
+    result = await Car.findById(req.params.id).populate('createdBy');
+  }
+
   if (!result) return next(new AppError(ERRORS.INVALID.NOT_FOUND, STATUS_CODE.NOT_FOUND));
+
   const ip = RequestIp.getClientIp(req);
-  //current user fav status
+
   if (!result.active || result.banned) {
     if (req.user) {
       const currentUser = req.user._id;
@@ -162,6 +179,26 @@ exports.getOne = catchAsync(async (req, res, next) => {
 });
 
 exports.updateOne = catchAsync(async (req, res, next) => {
+  let car, result, ObjectId;
+  ObjectId = mongoose.isValidObjectId(req.params.id);
+  if (ObjectId !== true) {
+    let stringValues = req.params.id;
+    let splitValues = stringValues.split('-');
+    // let idFromValues = splitValues.pop();
+    let idFromValues = splitValues.slice(-1)[0];
+    req.params.id = idFromValues;
+    let validId = mongoose.isValidObjectId(req.params.id);
+    if (validId === true) {
+      car = await Car.findById(req.params.id);
+    }
+  } else {
+    car = await Car.findById(req.params.id);
+  }
+
+  if (!car) {
+    return next(new AppError('No Result Found', STATUS_CODE.BAD_REQUEST));
+  }
+
   if (req.files.selectedImage) {
     let { Location } = await uploadS3(
       req.files.selectedImage[0],
@@ -172,6 +209,7 @@ exports.updateOne = catchAsync(async (req, res, next) => {
     );
 
     req.body.selectedImage = Location;
+    // when we only send selectedImage then it will push selectedImage to images array
     await Car.updateOne({ _id: req.params.id }, { $push: { image: Location } });
     var imagePath = Location;
   } else {
@@ -200,10 +238,13 @@ exports.updateOne = catchAsync(async (req, res, next) => {
   }
   if (req.body.image) {
     let array = [];
+    const selectedImage = car.selectedImage;
+    // if selectedImage's value is undefined
     if (imagePath === undefined) {
-      const car = await Car.findById(req.params.id);
-      const selectedImage = car.selectedImage;
-      array = [selectedImage];
+      // if selectedImage Field in collection is not undefined then do operation
+      if (selectedImage !== undefined) {
+        array = [selectedImage];
+      }
     } else {
       array = [imagePath];
     }
@@ -211,6 +252,7 @@ exports.updateOne = catchAsync(async (req, res, next) => {
       let images = req.body.image[i];
       array.push(images);
     }
+    // let unique = [...new Set(array)];
     req.body.image = array;
   }
 
@@ -230,10 +272,47 @@ exports.updateOne = catchAsync(async (req, res, next) => {
     return next(new AppError(ERRORS.UNAUTHORIZED.ASSOCIATED_PHONE, STATUS_CODE.UNAUTHORIZED));
   }
 
-  const result = await Car.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-    runValidators: true,
-  });
+  if (req.body.bodyColor) {
+    car.bodyColor = req.body.bodyColor;
+    await car.save();
+  }
+  if (req.body.make) {
+    car.make = req.body.make;
+    await car.save();
+  }
+  if (req.body.model) {
+    car.model = req.body.model;
+    await car.save();
+  }
+  if (req.body.city) {
+    car.city = req.body.city;
+    await car.save();
+  }
+  if (req.body.modelYear) {
+    car.modelYear = req.body.modelYear;
+    await car.save();
+  }
+
+  ObjectId = mongoose.isValidObjectId(req.params.id);
+  if (ObjectId !== true) {
+    let stringValues = req.params.id;
+    let splitValues = stringValues.split('-');
+    // let idFromValues = splitValues.pop();
+    let idFromValues = splitValues.slice(-1)[0];
+    req.params.id = idFromValues;
+    let validId = mongoose.isValidObjectId(req.params.id);
+    if (validId === true) {
+      result = await Car.findByIdAndUpdate(req.params.id, req.body, {
+        new: true,
+        runValidators: true,
+      });
+    }
+  } else {
+    result = await Car.findByIdAndUpdate(req.params.id, req.body, {
+      new: true,
+      runValidators: true,
+    });
+  }
 
   if (!result) return next(new AppError(ERRORS.INVALID.NOT_FOUND, STATUS_CODE.NOT_FOUND));
   res.status(STATUS_CODE.OK).json({
@@ -246,12 +325,27 @@ exports.updateOne = catchAsync(async (req, res, next) => {
 });
 
 exports.deleteOne = catchAsync(async (req, res, next) => {
-  const result = await Car.findByIdAndDelete(req.params.id);
+  let result;
+  const ObjectId = mongoose.isValidObjectId(req.params.id);
+  if (ObjectId !== true) {
+    let stringValues = req.params.id;
+    let splitValues = stringValues.split('-');
+    // let idFromValues = splitValues.pop();
+    let idFromValues = splitValues.slice(-1)[0];
+    req.params.id = idFromValues;
+
+    let validId = mongoose.isValidObjectId(req.params.id);
+    if (validId === true) {
+      result = await Car.findByIdAndDelete(req.params.id);
+    }
+  } else {
+    result = await Car.findByIdAndDelete(req.params.id);
+  }
+
   if (!result) return next(new AppError(ERRORS.INVALID.NOT_FOUND, STATUS_CODE.NOT_FOUND));
   res.status(STATUS_CODE.OK).json({
     status: STATUS.SUCCESS,
-    message: SUCCESS_MSG.SUCCESS_MESSAGES.OPERATION_SUCCESSFULL,
-    data: null,
+    message: SUCCESS_MSG.SUCCESS_MESSAGES.AD_DELETED,
   });
 });
